[SPARK-44691][SQL][CONNECT] Move Subclasses of AnalysisException to sql/api
### What changes were proposed in this pull request?

- Move subclasses of AnalysisException to sql/api
- Refactor some dependent utilities

### Why are the changes needed?

- Supporting throws these exceptions for better compatibility with the existing control flow behaviors

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

- Existing tests

Closes #42363 from heyihong/SPARK-44691.

Authored-by: Yihong He <yihong.he@databricks.com>
