+++ sql/api/src/main/java/org/apache/spark/sql/connector/catalog/IdentifierImpl.java
@@ -17,14 +17,14 @@
 
 package org.apache.spark.sql.connector.catalog;
 
+import org.apache.arrow.util.Preconditions;
+
 import java.util.Arrays;
 import java.util.Objects;
 import java.util.StringJoiner;
 
-import com.google.common.base.Preconditions;
-
 import org.apache.spark.annotation.Evolving;
-import org.apache.spark.sql.catalyst.util.package$;
+import org.apache.spark.sql.catalyst.util.QuotingUtils;
 
 /**
  *  An {@link Identifier} implementation.
@@ -56,9 +56,9 @@ class IdentifierImpl implements Identifier {
   public String toString() {
     StringJoiner joiner = new StringJoiner(".");
     for (String p : namespace) {
-      joiner.add(package$.MODULE$.quoteIfNeeded(p));
+      joiner.add(QuotingUtils.quoteIfNeeded(p));
     }
-    joiner.add(package$.MODULE$.quoteIfNeeded(name));
+    joiner.add(QuotingUtils.quoteIfNeeded(name));
     return joiner.toString();
   }
