+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/NoSuchItemException.scala
@@ -22,111 +22,7 @@ import org.apache.spark.sql.AnalysisException
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.catalog.CatalogTypes.TablePartitionSpec
 import org.apache.spark.sql.catalyst.util.{quoteIdentifier, quoteNameParts}
-import org.apache.spark.sql.connector.catalog.CatalogV2Implicits._
-import org.apache.spark.sql.connector.catalog.Identifier
 import org.apache.spark.sql.types.StructType
-
-
-/**
- * Thrown by a catalog when an item cannot be found. The analyzer will rethrow the exception
- * as an [[org.apache.spark.sql.AnalysisException]] with the correct position information.
- */
-case class NoSuchDatabaseException(db: String)
-  extends AnalysisException(errorClass = "SCHEMA_NOT_FOUND",
-    messageParameters = Map("schemaName" -> quoteIdentifier(db)))
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class NoSuchNamespaceException private(
-    message: String,
-    cause: Option[Throwable],
-    errorClass: Option[String],
-    messageParameters: Map[String, String])
-  extends AnalysisException(
-    message,
-    cause = cause,
-    errorClass = errorClass,
-    messageParameters = messageParameters) {
-
-  def this(errorClass: String, messageParameters: Map[String, String]) = {
-    this(
-      SparkThrowableHelper.getMessage(errorClass, messageParameters),
-      cause = None,
-      Some(errorClass),
-      messageParameters)
-  }
-
-  def this(namespace: Seq[String]) = {
-    this(errorClass = "SCHEMA_NOT_FOUND",
-      Map("schemaName" -> quoteNameParts(namespace)))
-  }
-
-  def this(namespace: Array[String]) = {
-    this(errorClass = "SCHEMA_NOT_FOUND",
-      Map("schemaName" -> quoteNameParts(namespace)))
-  }
-
-  def this(message: String, cause: Option[Throwable] = None) = {
-    this(
-      message,
-      cause,
-      errorClass = Some("SCHEMA_NOT_FOUND"),
-      messageParameters = Map.empty[String, String])
-  }
-}
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class NoSuchTableException private(
-    message: String,
-    cause: Option[Throwable],
-    errorClass: Option[String],
-    messageParameters: Map[String, String])
-  extends AnalysisException(
-    message,
-    cause = cause,
-    errorClass = errorClass,
-    messageParameters = messageParameters) {
-
-  def this(errorClass: String, messageParameters: Map[String, String]) = {
-    this(
-      SparkThrowableHelper.getMessage(errorClass, messageParameters),
-      cause = None,
-      Some(errorClass),
-      messageParameters)
-  }
-
-  def this(db: String, table: String) = {
-    this(errorClass = "TABLE_OR_VIEW_NOT_FOUND",
-      messageParameters = Map("relationName" ->
-        (quoteIdentifier(db) + "." + quoteIdentifier(table))))
-  }
-
-  def this(name : Seq[String]) = {
-    this(errorClass = "TABLE_OR_VIEW_NOT_FOUND",
-      messageParameters = Map("relationName" -> quoteNameParts(name)))
-  }
-
-  def this(tableIdent: Identifier) = {
-    this(errorClass = "TABLE_OR_VIEW_NOT_FOUND",
-      messageParameters = Map("relationName" -> tableIdent.quoted))
-  }
-
-  def this(message: String, cause: Option[Throwable] = None) = {
-    this(
-      message,
-      cause,
-      errorClass = Some("TABLE_OR_VIEW_NOT_FOUND"),
-      messageParameters = Map.empty[String, String])
-  }
-}
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class NoSuchViewException(errorClass: String, messageParameters: Map[String, String])
-  extends AnalysisException(errorClass, messageParameters) {
-
-  def this(ident: Identifier) =
-    this(errorClass = "VIEW_NOT_FOUND",
-      messageParameters = Map("relationName" -> ident.quoted))
-}
 
 // any changes to this class should be backward compatible as it may be used by external connectors
 class NoSuchPartitionException private(
@@ -169,48 +65,6 @@ class NoSuchPartitionException private(
   }
 }
 
-class NoSuchPermanentFunctionException(db: String, func: String)
-  extends AnalysisException(errorClass = "ROUTINE_NOT_FOUND",
-    Map("routineName" -> (quoteIdentifier(db) + "." + quoteIdentifier(func))))
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class NoSuchFunctionException private(
-    message: String,
-    cause: Option[Throwable],
-    errorClass: Option[String],
-    messageParameters: Map[String, String])
-  extends AnalysisException(
-    message,
-    cause = cause,
-    errorClass = errorClass,
-    messageParameters = messageParameters) {
-
-  def this(errorClass: String, messageParameters: Map[String, String]) = {
-    this(
-      SparkThrowableHelper.getMessage(errorClass, messageParameters),
-      cause = None,
-      Some(errorClass),
-      messageParameters)
-  }
-
-  def this(db: String, func: String) = {
-    this(errorClass = "ROUTINE_NOT_FOUND",
-      Map("routineName" -> (quoteIdentifier(db) + "." + quoteIdentifier(func))))
-  }
-
-  def this(identifier: Identifier) = {
-    this(errorClass = "ROUTINE_NOT_FOUND", Map("routineName" -> identifier.quoted))
-  }
-
-  def this(message: String, cause: Option[Throwable] = None) = {
-    this(
-      message,
-      cause,
-      errorClass = Some("ROUTINE_NOT_FOUND"),
-      messageParameters = Map.empty[String, String])
-  }
-}
-
 // any changes to this class should be backward compatible as it may be used by external connectors
 class NoSuchPartitionsException private(
     message: String,
@@ -249,45 +103,6 @@ class NoSuchPartitionsException private(
     this(
       message,
       errorClass = Some("PARTITIONS_NOT_FOUND"),
-      messageParameters = Map.empty[String, String])
-  }
-}
-
-class NoSuchTempFunctionException(func: String)
-  extends AnalysisException(errorClass = "ROUTINE_NOT_FOUND", Map("routineName" -> s"`$func`"))
-
-// any changes to this class should be backward compatible as it may be used by external connectors
-class NoSuchIndexException private(
-    message: String,
-    cause: Option[Throwable],
-    errorClass: Option[String],
-    messageParameters: Map[String, String])
-  extends AnalysisException(
-    message,
-    cause = cause,
-    errorClass = errorClass,
-    messageParameters = messageParameters) {
-
-  def this(
-      errorClass: String,
-      messageParameters: Map[String, String],
-      cause: Option[Throwable]) = {
-    this(
-      SparkThrowableHelper.getMessage(errorClass, messageParameters),
-      cause,
-      Some(errorClass),
-      messageParameters)
-  }
-
-  def this(indexName: String, tableName: String, cause: Option[Throwable]) = {
-    this("INDEX_NOT_FOUND", Map("indexName" -> indexName, "tableName" -> tableName), cause)
-  }
-
-  def this(message: String, cause: Option[Throwable] = None) = {
-    this(
-      message,
-      cause,
-      errorClass = Some("INDEX_NOT_FOUND"),
       messageParameters = Map.empty[String, String])
   }
 }
