+++ sql/catalyst/src/main/scala/org/apache/spark/sql/connector/catalog/CatalogV2Implicits.scala
@@ -24,7 +24,7 @@ import org.apache.spark.sql.catalyst.catalog.BucketSpec
 import org.apache.spark.sql.catalyst.expressions.AttributeReference
 import org.apache.spark.sql.catalyst.parser.CatalystSqlParser
 import org.apache.spark.sql.catalyst.types.DataTypeUtils
-import org.apache.spark.sql.catalyst.util.quoteIfNeeded
+import org.apache.spark.sql.catalyst.util.{quoteIfNeeded, QuotingUtils}
 import org.apache.spark.sql.connector.expressions.{BucketTransform, FieldReference, IdentityTransform, LogicalExpressions, Transform}
 import org.apache.spark.sql.errors.{QueryCompilationErrors, QueryExecutionErrors}
 import org.apache.spark.sql.types.StructType
@@ -110,7 +110,7 @@ private[sql] object CatalogV2Implicits {
   }
 
   implicit class NamespaceHelper(namespace: Array[String]) {
-    def quoted: String = namespace.map(quoteIfNeeded).mkString(".")
+    def quoted: String = QuotingUtils.quoted(namespace)
   }
 
   implicit class FunctionIdentifierHelper(ident: FunctionIdentifier) {
@@ -126,11 +126,7 @@ private[sql] object CatalogV2Implicits {
 
   implicit class IdentifierHelper(ident: Identifier) {
     def quoted: String = {
-      if (ident.namespace.nonEmpty) {
-        ident.namespace.map(quoteIfNeeded).mkString(".") + "." + quoteIfNeeded(ident.name)
-      } else {
-        quoteIfNeeded(ident.name)
-      }
+      QuotingUtils.quoted(ident)
     }
 
     def original: String = ident.namespace() :+ ident.name() mkString "."
